r6484/sofa-dev : FIX : removing a few more warnings
 //partial sync (2/3 changes)//

+++ modules/sofa/component/constraint/LCPConstraintSolver.cpp
@@ -1545,8 +1545,8 @@ void LCPConstraintSolver::draw()
                 int colid = (level * 12 + ((int)level < merge_local_levels ? (cb % 2) : 0)) % 72;
                 color.i = colors[colid + 0];
                 simulation::getSimulation()->DrawUtility.drawArrow(
-                    centerFine,centerFine+dirFineN*radius*2,
-                    radius*2*0.03,
+                    centerFine,centerFine+dirFineN*radius*2.0f,
+                    (float)radius*2.0f*0.03f,
                     defaulttype::Vec<4,float>((float)(color.b[0]) * (1.0f/255.0f),
                             (float)(color.b[1]) * (1.0f/255.0f),
                             (float)(color.b[2]) * (1.0f/255.0f),
@@ -1556,7 +1556,7 @@ void LCPConstraintSolver::draw()
                     color.i = colors[colid + 2];
                     simulation::getSimulation()->DrawUtility.drawArrow(
                         centerFine-dirFineT1*radius*_mu,centerFine+dirFineT1*radius*_mu,
-                        radius*_mu*0.03,
+                        (float)(radius*_mu*0.03f),
                         defaulttype::Vec<4,float>((float)(color.b[0]) * (1.0f/255.0f),
                                 (float)(color.b[1]) * (1.0f/255.0f),
                                 (float)(color.b[2]) * (1.0f/255.0f),
@@ -1564,7 +1564,7 @@ void LCPConstraintSolver::draw()
                     color.i = colors[colid + 4];
                     simulation::getSimulation()->DrawUtility.drawArrow(
                         centerFine-dirFineT2*radius*_mu,centerFine+dirFineT2*radius*_mu,
-                        radius*_mu*0.03,
+                        (float)(radius*_mu*0.03f),
                         defaulttype::Vec<4,float>(color.b[0] * (1.0f/255.0f),
                                 color.b[1] * (1.0f/255.0f),
                                 color.b[2] * (1.0f/255.0f),
